Rename ray._private.structured_logging to ray._private.ray_logging
