+++ python/ray/tests/BUILD
@@ -154,6 +154,7 @@ py_test_module_list(
     "accelerators/test_amd_gpu.py",
     "test_log_dedup.py",
     "test_logging.py",
+    "test_logging_2.py",
     "test_memory_scheduling.py",
     "test_nested_task.py",
     "test_metrics.py",
@@ -190,7 +191,6 @@ py_test_module_list(
     "test_shuffle.py",
     "test_state_api_log.py",
     "test_state_api_summary.py",
-    "test_structured_logging.py",
     "test_tempfile.py",
     "test_tensorflow.py",
     "test_tls_auth.py",
